Link property multiselect

+++ src/main/scala/fi/liikennevirasto/digiroad2/Digiroad2Api.scala
@@ -1,19 +1,18 @@
 package fi.liikennevirasto.digiroad2
 
-import fi.liikennevirasto.digiroad2.asset.oracle.AssetPropertyConfiguration
-import fi.liikennevirasto.digiroad2.linearasset.{NewLimit, LinearAssetProvider}
-import org.scalatra._
-import org.json4s._
-import org.scalatra.json._
-import fi.liikennevirasto.digiroad2.Digiroad2Context._
-import fi.liikennevirasto.digiroad2.asset._
-import org.joda.time.{LocalDate, DateTime}
-import fi.liikennevirasto.digiroad2.authentication.{UserNotFoundException, RequestHeaderAuthentication, UnauthenticatedException}
-import org.slf4j.LoggerFactory
-import fi.liikennevirasto.digiroad2.asset.BoundingRectangle
-import fi.liikennevirasto.digiroad2.user.{User}
-import org.apache.commons.lang3.StringUtils.isBlank
 import com.newrelic.api.agent.NewRelic
+import fi.liikennevirasto.digiroad2.Digiroad2Context._
+import fi.liikennevirasto.digiroad2.asset.{BoundingRectangle, _}
+import fi.liikennevirasto.digiroad2.asset.oracle.AssetPropertyConfiguration
+import fi.liikennevirasto.digiroad2.authentication.{RequestHeaderAuthentication, UnauthenticatedException, UserNotFoundException}
+import fi.liikennevirasto.digiroad2.linearasset.{LinearAssetProvider, NewLimit, RoadLinkPartitioner}
+import fi.liikennevirasto.digiroad2.user.User
+import org.apache.commons.lang3.StringUtils.isBlank
+import org.joda.time.DateTime
+import org.json4s._
+import org.scalatra._
+import org.scalatra.json._
+import org.slf4j.LoggerFactory
 
 class Digiroad2Api(val roadLinkService: RoadLinkService,
                    val linearAssetProvider: LinearAssetProvider,
@@ -36,8 +35,7 @@ with GZipSupport {
     contentType = formats("json") + "; charset=utf-8"
     try {
       authenticateForApi(request)(userProvider)
-      if(request.isWrite
-        && userProvider.getCurrentUser().hasWriteAccess() == false){
+      if(request.isWrite && !userProvider.getCurrentUser().hasWriteAccess()){
         halt(Unauthorized("No write permissions"))
       }
     } catch {
@@ -56,7 +54,7 @@ with GZipSupport {
   }
 
   get("/massTransitStops") {
-    val user = userProvider.getCurrentUser
+    val user = userProvider.getCurrentUser()
     val bbox = params.get("bbox").map(constructBoundingRectangle).getOrElse(halt(BadRequest("Bounding box was missing")))
     validateBoundingBox(bbox)
     useVVHGeometry match {
@@ -230,12 +228,12 @@ with GZipSupport {
     }
   }
 
-  private def getRoadLinksFromVVH(municipalities: Set[Int])(bbox: String): Seq[Map[String, Any]]  = {
+  private def getRoadLinksFromVVH(municipalities: Set[Int])(bbox: String): Seq[Seq[Map[String, Any]]]  = {
     val boundingRectangle = constructBoundingRectangle(bbox)
     validateBoundingBox(boundingRectangle)
-    roadLinkService.getRoadLinksFromVVH(
-      bounds = boundingRectangle,
-      municipalities = municipalities).map { roadLink =>
+    val roadLinks = roadLinkService.getRoadLinksFromVVH(boundingRectangle, municipalities)
+    val partitionedRoadLinks = RoadLinkPartitioner.partition(roadLinks)
+    partitionedRoadLinks.map { group => group.map { roadLink =>
       Map(
         "mmlId" -> roadLink.mmlId,
         "points" -> roadLink.geometry,
@@ -253,7 +251,7 @@ with GZipSupport {
         "maxAddressNumberRight" -> roadLink.attributes.get("MAXANRIGHT"),
         "minAddressNumberLeft" -> roadLink.attributes.get("MINANLEFT"),
         "maxAddressNumberLeft" -> roadLink.attributes.get("MAXANLEFT"))
-    }
+    } }
   }
 
   get("/roadlinks") {
@@ -304,24 +302,30 @@ with GZipSupport {
     roadLinkService.getIncompleteLinks(includedMunicipalities)
   }
 
-  put("/linkproperties/:id") {
-    val id = params("id").toLong
-    val user = userProvider.getCurrentUser()
+  put("/linkproperties") {
+    val ids = (parsedBody \ "mmlIds").extract[Seq[Long]]
+    updateRoadLinkProperties(ids)
+  }
+
+  def updateRoadLinkProperties(ids: Seq[Long]): Seq[Map[String, Any]] = {
     val trafficDirection = TrafficDirection((parsedBody \ "trafficDirection").extract[String])
     val functionalClass = (parsedBody \ "functionalClass").extract[Int]
     val linkType = LinkType((parsedBody \ "linkType").extract[Int])
-    def municipalityValidation(municipalityCode: Int) = hasWriteAccess(user, municipalityCode)
 
-    roadLinkService.updateProperties(id, functionalClass, linkType, trafficDirection, user.username, municipalityValidation).map { roadLink =>
-      Map("mmlId" -> roadLink.mmlId,
-        "points" -> roadLink.geometry,
-        "administrativeClass" -> roadLink.administrativeClass.toString,
-        "functionalClass" -> roadLink.functionalClass,
-        "trafficDirection" -> roadLink.trafficDirection.toString,
-        "modifiedAt" -> roadLink.modifiedAt,
-        "modifiedBy" -> roadLink.modifiedBy,
-        "linkType" -> roadLink.linkType.value)
-    }.getOrElse(NotFound("Road link with MML ID " + id + " not found"))
+    val user = userProvider.getCurrentUser()
+    def municipalityValidation(municipalityCode: Int) = validateUserMunicipalityAccess(user)(municipalityCode)
+    ids.map { id =>
+      roadLinkService.updateProperties(id, functionalClass, linkType, trafficDirection, user.username, municipalityValidation).map { roadLink =>
+        Map("mmlId" -> roadLink.mmlId,
+          "points" -> roadLink.geometry,
+          "administrativeClass" -> roadLink.administrativeClass.toString,
+          "functionalClass" -> roadLink.functionalClass,
+          "trafficDirection" -> roadLink.trafficDirection.toString,
+          "modifiedAt" -> roadLink.modifiedAt,
+          "modifiedBy" -> roadLink.modifiedBy,
+          "linkType" -> roadLink.linkType.value)
+      }.getOrElse(halt(NotFound("Road link with MML ID " + id + " not found")))
+    }
   }
 
   get("/assetTypeProperties/:assetTypeId") {
@@ -342,11 +346,10 @@ with GZipSupport {
     case ise: IllegalStateException => halt(InternalServerError("Illegal state: " + ise.getMessage))
     case ue: UnauthenticatedException => halt(Unauthorized("Not authenticated"))
     case unf: UserNotFoundException => halt(Forbidden(unf.username))
-    case e: Exception => {
+    case e: Exception =>
       logger.error("API Error", e)
       NewRelic.noticeError(e)
       halt(InternalServerError("API error"))
-    }
   }
 
   private def validateBoundingBox(bbox: BoundingRectangle): Unit = {
@@ -379,7 +382,7 @@ with GZipSupport {
             "sideCode" -> link.sideCode,
             "trafficDirection" -> link.trafficDirection,
             "value" -> link.value,
-            "points" -> link.points,
+            "points" -> link.geometry,
             "startMeasure" -> link.startMeasure,
             "endMeasure" -> link.endMeasure,
             "modifiedBy" -> link.modifiedBy,
@@ -435,7 +438,7 @@ with GZipSupport {
   put("/numericallimits/:id") {
     val user = userProvider.getCurrentUser()
     val id = params("id").toLong
-    if (!user.hasEarlyAccess() || !assetService.getMunicipalityCodes(id).forall(user.isAuthorizedToWrite(_))) {
+    if (!user.hasEarlyAccess() || !assetService.getMunicipalityCodes(id).forall(user.isAuthorizedToWrite)) {
       halt(Unauthorized("User not authorized"))
     }
     val expiredOption: Option[Boolean] = (parsedBody \ "expired").extractOpt[Boolean]
@@ -455,14 +458,14 @@ with GZipSupport {
     val user = userProvider.getCurrentUser()
     val roadLinkId = (parsedBody \ "roadLinkId").extract[Long]
     val municipalityCode = RoadLinkService.getMunicipalityCode(roadLinkId)
-    hasWriteAccess(user, municipalityCode.get)
+    validateUserMunicipalityAccess(user)(municipalityCode.get)
     val typeId = params.getOrElse("typeId", halt(BadRequest("Missing mandatory 'typeId' parameter"))).toInt
     val value = (parsedBody \ "value").extract[BigInt]
     validateNumericalLimitValue(value)
     val username = user.username
     NumericalLimitService.createNumericalLimit(typeId = typeId,
                                          roadLinkId = roadLinkId,
-                                         value = value.intValue,
+                                         value = value.intValue(),
                                          username = username)
   }
 
@@ -470,14 +473,14 @@ with GZipSupport {
     val user = userProvider.getCurrentUser()
     val roadLinkId = (parsedBody \ "roadLinkId").extract[Long]
     val municipalityCode = RoadLinkService.getMunicipalityCode(roadLinkId)
-    hasWriteAccess(user, municipalityCode.get)
+    validateUserMunicipalityAccess(user)(municipalityCode.get)
     val value = (parsedBody \ "value").extract[BigInt]
     validateNumericalLimitValue(value)
     val expired = (parsedBody \ "expired").extract[Boolean]
     val id = params("id").toLong
     val username = user.username
     val measure = (parsedBody \ "splitMeasure").extract[Double]
-    NumericalLimitService.split(id, roadLinkId, measure, value.intValue, expired, username)
+    NumericalLimitService.split(id, roadLinkId, measure, value.intValue(), expired, username)
   }
 
   put("/speedlimits") {
@@ -502,7 +505,7 @@ with GZipSupport {
       (parsedBody \ "existingValue").extract[Int],
       (parsedBody \ "createdValue").extract[Int],
       user.username,
-      validateUserMunicipalityAccess(user)_)
+      validateUserMunicipalityAccess(user))
   }
 
   post("/speedlimits/:speedLimitId/separate") {
@@ -537,12 +540,6 @@ with GZipSupport {
     }
   }
 
-  def hasWriteAccess(user: User, municipality: Int) {
-    if (!user.hasEarlyAccess() || !user.isAuthorizedToWrite(municipality)) {
-      halt(Unauthorized("User not authorized"))
-    }
-  }
-
   get("/manoeuvres") {
     val user = userProvider.getCurrentUser()
     val municipalities: Set[Int] = if (user.isOperator()) Set() else user.configuration.authorizedMunicipalities
@@ -562,7 +559,7 @@ with GZipSupport {
 
     val manoeuvreIds = manoeuvres.map { manoeuvre =>
       val municipality = RoadLinkService.getMunicipalityCode(manoeuvre.sourceRoadLinkId)
-      hasWriteAccess(user, municipality.get)
+      validateUserMunicipalityAccess(user)(municipality.get)
       ManoeuvreService.createManoeuvre(user.username, manoeuvre)
     }
     Created(manoeuvreIds)
@@ -575,7 +572,7 @@ with GZipSupport {
 
     manoeuvreIds.foreach { manoeuvreId =>
       val sourceRoadLinkId = ManoeuvreService.getSourceRoadLinkIdById(manoeuvreId)
-      hasWriteAccess(user, RoadLinkService.getMunicipalityCode(sourceRoadLinkId).get)
+      validateUserMunicipalityAccess(user)(RoadLinkService.getMunicipalityCode(sourceRoadLinkId).get)
       ManoeuvreService.deleteManoeuvre(user.username, manoeuvreId)
     }
   }
@@ -588,7 +585,7 @@ with GZipSupport {
       .map{case(id, updates) => (id.toLong, updates)}
     manoeuvreUpdates.foreach{ case(id, updates) =>
       val sourceRoadLinkId = ManoeuvreService.getSourceRoadLinkIdById(id)
-      hasWriteAccess(user, RoadLinkService.getMunicipalityCode(sourceRoadLinkId).get)
+      validateUserMunicipalityAccess(user)(RoadLinkService.getMunicipalityCode(sourceRoadLinkId).get)
       ManoeuvreService.updateManoeuvre(user.username, id, updates)
     }
   }
